+++ internal/depgraph/deps_mod.go
@@ -8,25 +8,27 @@ import (
 
 	"go.uber.org/zap"
 
+	"github.com/Helcaraxan/gomod/internal/logger"
 	"github.com/Helcaraxan/gomod/internal/util"
 )
 
-func (g *DepGraph) overlayModuleDependencies() error {
-	g.log.Debug("Overlaying module-based dependency information over the import dependency graph.")
+func (g *DepGraph) overlayModuleDependencies(dl *logger.Builder) error {
+	log := dl.Domain(logger.ModuleDependencyDomain)
+	log.Debug("Overlaying module-based dependency information over the import dependency graph.")
 
-	raw, _, err := util.RunCommand(g.log, g.Main.Info.Dir, "go", "mod", "graph")
+	raw, _, err := util.RunCommand(log, g.Main.Info.Dir, "go", "mod", "graph")
 	if err != nil {
 		return err
 	}
 
 	for _, depString := range strings.Split(strings.TrimSpace(string(raw)), "\n") {
-		g.log.Debug("Parsing dependency", zap.String("reference", depString))
-		modDep, ok := g.parseDependency(depString)
+		log.Debug("Parsing dependency", zap.String("reference", depString))
+		modDep, ok := g.parseDependency(log, depString)
 		if !ok {
 			continue
 		}
 
-		g.log.Debug(
+		log.Debug(
 			"Overlaying module dependency.",
 			zap.String("version", modDep.targetVersion),
 			zap.String("source", modDep.source.Name()),
@@ -42,7 +44,7 @@ func (g *DepGraph) overlayModuleDependencies() error {
 		}
 	}
 
-	if err := g.markIndirects(); err != nil {
+	if err := g.markIndirects(log); err != nil {
 		return err
 	}
 
@@ -56,10 +58,10 @@ type moduleDependency struct {
 	targetVersion string
 }
 
-func (g *DepGraph) parseDependency(depString string) (*moduleDependency, bool) {
+func (g *DepGraph) parseDependency(log *logger.Logger, depString string) (*moduleDependency, bool) {
 	depContent := depRE.FindStringSubmatch(depString)
 	if len(depContent) == 0 {
-		g.log.Warn("Skipping ill-formed line in 'go mod graph' output.", zap.String("line", depString))
+		log.Warn("Skipping ill-formed line in 'go mod graph' output.", zap.String("line", depString))
 		return nil, false
 	}
 
@@ -68,18 +70,18 @@ func (g *DepGraph) parseDependency(depString string) (*moduleDependency, bool) {
 
 	source, ok := g.getModule(sourceName)
 	if !ok {
-		g.log.Warn("Encountered a dependency edge starting at an unknown module.", zap.String("source", sourceName), zap.String("target", targetName))
+		log.Warn("Encountered a dependency edge starting at an unknown module.", zap.String("source", sourceName), zap.String("target", targetName))
 		return nil, false
 	}
 	target, ok := g.getModule(targetName)
 	if !ok {
-		g.log.Warn("Encountered a dependency edge ending at an unknown module.", zap.String("source", sourceName), zap.String("target", targetName))
+		log.Warn("Encountered a dependency edge ending at an unknown module.", zap.String("source", sourceName), zap.String("target", targetName))
 		return nil, false
 
 	}
 
 	if sourceVersion != source.Info.Version {
-		g.log.Debug(
+		log.Debug(
 			"Skipping edge as we are not using the specified source version.",
 			zap.String("source", sourceName),
 			zap.String("version", sourceVersion),
@@ -87,7 +89,7 @@ func (g *DepGraph) parseDependency(depString string) (*moduleDependency, bool) {
 		)
 		return nil, false
 	}
-	g.log.Debug(
+	log.Debug(
 		"Recording module dependency.",
 		zap.String("source", sourceName),
 		zap.String("version", sourceVersion),
@@ -102,11 +104,11 @@ func (g *DepGraph) parseDependency(depString string) (*moduleDependency, bool) {
 	}, true
 }
 
-func (g *DepGraph) markIndirects() error {
+func (g *DepGraph) markIndirects(log *logger.Logger) error {
 	for _, node := range g.Graph.GetLevel(int(LevelModules)).List() {
 		module := node.(*Module)
 
-		log := g.log.With(zap.String("module", module.Name()))
+		log := log.With(zap.String("module", module.Name()))
 		log.Debug("Finding indirect dependencies for module.")
 
 		if module.Info.GoMod == "" {
@@ -116,14 +118,14 @@ func (g *DepGraph) markIndirects() error {
 
 		modContent, err := ioutil.ReadFile(module.Info.GoMod)
 		if err != nil {
-			g.log.Error("Failed to read content of go.mod file.", zap.String("path", module.Info.GoMod), zap.Error(err))
+			log.Error("Failed to read content of go.mod file.", zap.String("path", module.Info.GoMod), zap.Error(err))
 			return err
 		}
 
 		indirectDepRE := regexp.MustCompile(`^	([^\s]+) [^\s]+ // indirect$`)
 		for _, line := range bytes.Split(modContent, []byte("\n")) {
 			if m := indirectDepRE.FindSubmatch(line); len(m) == 2 {
-				g.log.Debug("Found indirect dependency.", zap.String("consumer", module.Name()), zap.String("dependency", string(m[1])))
+				log.Debug("Found indirect dependency.", zap.String("consumer", module.Name()), zap.String("dependency", string(m[1])))
 				module.Indirects[string(m[1])] = true
 			}
 		}